Turned Double, Int, Long and Float objects into static private classes in NumberObject so they are no longer acessible directly
